requirements.txt
Django==4.0.6
djangorestframework==3.13.1
django-debug-toolbar==3.5.0
redis==4.3.4
django-redis==5.2.0
drf-yasg==1.21.0
python-dotenv==0.20.0
django-filter==22.1
